Dodana prijava i uzimanje tabela iz baze

+++ Models/Product.cs
@@ -14,7 +14,7 @@ namespace WpfApp1.Models
         public string Name {
             get; set;
         }
-        public double Price {
+        public decimal Price {
             get;set;
         }
         public string Description
